requirements.txt
requests==2.32.3
prometheus_client==0.20.0
PyYAML==6.0.1
